Pull related classes out of `WebSocketTransport` file and into their own files
